Workaround, issue 12
https://github.com/moltin/laravel-cart/issues/12
