+++ composer.json
@@ -10,7 +10,7 @@
     "minimum-stability": "dev",
     "require": {
         "php": ">=5.3.0",
-        "laravel/framework": "4.*",
+        "laravel/framework": "4.1.*",
         "moltin/cart": "dev-master"
     },
     "require-dev": {
