+++ src/Moltin/Cart/Identifier/RequestCookie.php
@@ -4,7 +4,7 @@
  * @package Moltin\Cart\Storage
  * @author Theo den Hollander <theo@hollandware.com>
  */
-use Illuminate\Support\Facades\Cache;
+
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Input;
